+++ test/plotje_book/pose_rules_generated_test.clj
@@ -21,7 +21,7 @@
    (:layers pose)
    (update
     :layers
-    (partial mapv (fn* [p1__78072#] (dissoc p1__78072# :data))))
+    (partial mapv (fn* [p1__11193#] (dissoc p1__11193# :data))))
    (:poses pose)
    (update :poses (partial mapv strip-data)))))
 
@@ -35,21 +35,21 @@
   (kind/pprint (strip-data pose))))
 
 
-(def v8_l85 (-> iris (pj/pose :sepal-length :sepal-width)))
+(def v8_l92 (-> iris (pj/pose :sepal-length :sepal-width)))
 
 
 (deftest
- t9_l88
- (is ((fn [v] (= 150 (:points (pj/svg-summary v)))) v8_l85)))
+ t9_l95
+ (is ((fn [v] (= 150 (:points (pj/svg-summary v)))) v8_l92)))
 
 
 (def
- v10_l90
+ v10_l97
  (-> iris (pj/pose :sepal-length :sepal-width) pose-summary))
 
 
 (deftest
- t11_l94
+ t11_l101
  (is
   ((fn
     [pose]
@@ -57,39 +57,39 @@
      (= {:x :sepal-length, :y :sepal-width} (:mapping pose))
      (= [] (:layers pose))
      (not (contains? pose :poses))))
-   v10_l90)))
+   v10_l97)))
 
 
-(def v13_l104 (-> iris (pj/pose {:color :species}) pose-summary))
+(def v13_l111 (-> iris (pj/pose {:color :species}) pose-summary))
 
 
 (deftest
- t14_l108
+ t14_l115
  (is
   ((fn
     [pose]
     (and
      (= {:color :species} (:mapping pose))
      (not (contains? pose :poses))))
-   v13_l104)))
+   v13_l111)))
 
 
-(def v16_l120 (-> iris pj/pose (pj/pose :sepal-length :sepal-width)))
+(def v16_l127 (-> iris pj/pose (pj/pose :sepal-length :sepal-width)))
 
 
 (deftest
- t17_l124
+ t17_l131
  (is
   ((fn
     [pose]
     (and
      (= {:x :sepal-length, :y :sepal-width} (:mapping pose))
      (not (contains? pose :poses))))
-   v16_l120)))
+   v16_l127)))
 
 
 (def
- v19_l131
+ v19_l138
  (->
   iris
   (pj/pose {:color :species})
@@ -97,18 +97,18 @@
 
 
 (deftest
- t20_l135
+ t20_l142
  (is
   ((fn
     [pose]
     (=
      {:x :sepal-length, :y :sepal-width, :color :species}
      (:mapping pose)))
-   v19_l131)))
+   v19_l138)))
 
 
 (def
- v22_l143
+ v22_l150
  (=
   (->
    iris
@@ -118,11 +118,11 @@
   (pj/pose iris :sepal-length :sepal-width {:color :species})))
 
 
-(deftest t23_l149 (is (true? v22_l143)))
+(deftest t23_l156 (is (true? v22_l150)))
 
 
 (def
- v25_l160
+ v25_l167
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -130,7 +130,7 @@
 
 
 (deftest
- t26_l164
+ t26_l171
  (is
   ((fn
     [pose]
@@ -142,11 +142,11 @@
      (=
       {:x :petal-length, :y :petal-width}
       (:mapping (second (:poses pose))))))
-   v25_l160)))
+   v25_l167)))
 
 
 (def
- v28_l176
+ v28_l183
  (->
   iris
   (pj/pose :sepal-length :sepal-width {:color :species})
@@ -154,7 +154,7 @@
 
 
 (def
- v29_l180
+ v29_l187
  (->
   iris
   (pj/pose :sepal-length :sepal-width {:color :species})
@@ -163,7 +163,7 @@
 
 
 (deftest
- t30_l185
+ t30_l192
  (is
   ((fn
     [pose]
@@ -175,11 +175,11 @@
      (=
       {:x :petal-length, :y :petal-width}
       (:mapping (second (:poses pose))))))
-   v29_l180)))
+   v29_l187)))
 
 
 (def
- v32_l197
+ v32_l204
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -188,7 +188,7 @@
 
 
 (def
- v34_l205
+ v34_l212
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -198,18 +198,18 @@
 
 
 (deftest
- t35_l211
+ t35_l218
  (is
   ((fn
     [pose]
     (and
      (= "Iris" (get-in pose [:opts :title]))
      (not (contains? (first (:poses pose)) :opts))))
-   v34_l205)))
+   v34_l212)))
 
 
 (def
- v37_l224
+ v37_l231
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -217,7 +217,7 @@
 
 
 (deftest
- t38_l228
+ t38_l235
  (is
   ((fn
     [pose]
@@ -227,11 +227,11 @@
      (=
       {:x :sepal-length, :y :sepal-width}
       (:mapping (first (:poses pose))))))
-   v37_l224)))
+   v37_l231)))
 
 
 (def
- v40_l241
+ v40_l248
  (=
   (->
    iris
@@ -244,11 +244,11 @@
    (pj/pose :petal-length :petal-width))))
 
 
-(deftest t41_l249 (is (true? v40_l241)))
+(deftest t41_l256 (is (true? v40_l248)))
 
 
 (def
- v43_l260
+ v43_l267
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -257,7 +257,7 @@
 
 
 (deftest
- t44_l265
+ t44_l272
  (is
   ((fn
     [pose]
@@ -267,11 +267,11 @@
      (= 2 (count (:poses pose)))
      (= [] (:layers (first (:poses pose))))
      (= [] (:layers (second (:poses pose))))))
-   v43_l260)))
+   v43_l267)))
 
 
 (def
- v46_l278
+ v46_l285
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -280,7 +280,7 @@
 
 
 (deftest
- t47_l283
+ t47_l290
  (is
   ((fn
     [pose]
@@ -288,11 +288,11 @@
      (or (not (contains? pose :layers)) (= [] (:layers pose)))
      (= 1 (count (:layers (first (:poses pose)))))
      (= [] (:layers (second (:poses pose))))))
-   v46_l278)))
+   v46_l285)))
 
 
 (def
- v49_l299
+ v49_l306
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -301,7 +301,7 @@
 
 
 (deftest
- t50_l304
+ t50_l311
  (is
   ((fn
     [pose]
@@ -312,11 +312,11 @@
        {:x :petal-length, :y :petal-width}
        {:x :sepal-length, :y :petal-length}]
       (mapv :mapping (:poses pose)))))
-   v49_l299)))
+   v49_l306)))
 
 
 (def
- v52_l315
+ v52_l322
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -325,7 +325,7 @@
 
 
 (def
- v53_l320
+ v53_l327
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -335,7 +335,7 @@
 
 
 (deftest
- t54_l326
+ t54_l333
  (is
   ((fn
     [pose]
@@ -345,28 +345,28 @@
      (=
       {:x :sepal-length, :y :sepal-width}
       (:mapping (first (:poses pose))))))
-   v53_l320)))
+   v53_l327)))
 
 
 (def
- v56_l341
+ v56_l348
  (def leaf-pose (-> iris (pj/pose :sepal-length :sepal-width))))
 
 
-(def v57_l343 leaf-pose)
+(def v57_l350 leaf-pose)
 
 
-(def v58_l345 (pose-summary leaf-pose))
+(def v58_l352 (pose-summary leaf-pose))
 
 
-(def v60_l349 (= leaf-pose (pj/pose leaf-pose)))
+(def v60_l356 (= leaf-pose (pj/pose leaf-pose)))
 
 
-(deftest t61_l351 (is (true? v60_l349)))
+(deftest t61_l358 (is (true? v60_l356)))
 
 
 (def
- v63_l355
+ v63_l362
  (def
   composite-pose
   (->
@@ -375,27 +375,27 @@
    (pj/pose :petal-length :petal-width))))
 
 
-(def v64_l360 composite-pose)
+(def v64_l367 composite-pose)
 
 
-(def v65_l362 (pose-summary composite-pose))
+(def v65_l369 (pose-summary composite-pose))
 
 
-(def v67_l366 (= composite-pose (pj/pose composite-pose)))
+(def v67_l373 (= composite-pose (pj/pose composite-pose)))
 
 
-(deftest t68_l368 (is (true? v67_l366)))
+(deftest t68_l375 (is (true? v67_l373)))
 
 
 (def
- v70_l377
+ v70_l384
  (pj/arrange
   [(-> iris (pj/pose :sepal-length :sepal-width) pj/lay-point)
    (-> iris (pj/pose :petal-length :petal-width) pj/lay-point)]))
 
 
 (deftest
- t71_l381
+ t71_l388
  (is
   ((fn
     [pose]
@@ -404,11 +404,11 @@
      (= :vertical (get-in pose [:layout :direction]))
      (= 1 (count (:poses pose)))
      (= 2 (count (:poses (first (:poses pose)))))))
-   v70_l377)))
+   v70_l384)))
 
 
 (def
- v73_l393
+ v73_l400
  (pj/arrange
   [(pj/pose iris :sepal-length :sepal-width)
    (pj/pose iris :petal-length :petal-width)]
@@ -416,18 +416,18 @@
 
 
 (deftest
- t74_l399
+ t74_l406
  (is
   ((fn
     [pose]
     (and
      (= "Arranged" (get-in pose [:opts :title]))
      (= #{:y} (get-in pose [:opts :share-scales]))))
-   v73_l393)))
+   v73_l400)))
 
 
 (def
- v76_l415
+ v76_l422
  (->
   iris
   (pj/pose
@@ -436,7 +436,7 @@
 
 
 (deftest
- t77_l420
+ t77_l427
  (is
   ((fn
     [pose]
@@ -444,13 +444,13 @@
      (= {:color :species} (:mapping pose))
      (= 2 (count (:poses pose)))
      (every?
-      (fn* [p1__78073#] (= 2 (count (:poses p1__78073#))))
+      (fn* [p1__11194#] (= 2 (count (:poses p1__11194#))))
       (:poses pose))))
-   v76_l415)))
+   v76_l422)))
 
 
 (def
- v79_l429
+ v79_l436
  (let
   [a
    (->
@@ -471,16 +471,16 @@
   (= a b)))
 
 
-(deftest t80_l439 (is (true? v79_l429)))
+(deftest t80_l446 (is (true? v79_l436)))
 
 
 (def
- v82_l465
+ v82_l472
  (-> iris (pj/pose :sepal-length :sepal-width) pj/lay-point))
 
 
 (deftest
- t83_l469
+ t83_l476
  (is
   ((fn
     [pose]
@@ -488,11 +488,11 @@
      (= 1 (count (:layers pose)))
      (= :point (:layer-type (first (:layers pose))))
      (empty? (or (:mapping (first (:layers pose))) {}))))
-   v82_l465)))
+   v82_l472)))
 
 
 (def
- v85_l478
+ v85_l485
  (->
   (pj/arrange
    [(pj/pose iris :sepal-length :sepal-width)
@@ -501,7 +501,7 @@
 
 
 (def
- v86_l483
+ v86_l490
  (->
   (pj/arrange
    [(pj/pose iris :sepal-length :sepal-width)
@@ -511,7 +511,7 @@
 
 
 (deftest
- t87_l489
+ t87_l496
  (is
   ((fn
     [pose]
@@ -519,11 +519,11 @@
      (contains? pose :poses)
      (= 1 (count (:layers pose)))
      (= :point (:layer-type (first (:layers pose))))))
-   v86_l483)))
+   v86_l490)))
 
 
 (def
- v89_l501
+ v89_l508
  (let
   [before
    (pj/arrange
@@ -538,11 +538,11 @@
   [(count (or (:layers before) [])) (count (or (:layers after) []))]))
 
 
-(deftest t90_l511 (is ((fn [counts] (= [0 1] counts)) v89_l501)))
+(deftest t90_l518 (is ((fn [counts] (= [0 1] counts)) v89_l508)))
 
 
 (def
- v92_l526
+ v92_l533
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -551,7 +551,7 @@
 
 
 (deftest
- t93_l531
+ t93_l538
  (is
   ((fn
     [pose]
@@ -560,11 +560,11 @@
      (= 1 (count (:layers (first (:poses pose)))))
      (= 0 (count (:layers (second (:poses pose)))))
      (= :point (:layer-type (first (:layers (first (:poses pose))))))))
-   v92_l526)))
+   v92_l533)))
 
 
 (def
- v95_l548
+ v95_l555
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -572,7 +572,7 @@
 
 
 (deftest
- t96_l552
+ t96_l559
  (is
   ((fn
     [fr]
@@ -584,11 +584,11 @@
      (=
       {:x :petal-length, :y :petal-width}
       (:mapping (second (:poses fr))))))
-   v95_l548)))
+   v95_l555)))
 
 
 (def
- v98_l573
+ v98_l579
  (->
   {:X [1 2 3 4 5], :Y [1 2 3 4 5], :Z [1 4 9 16 25]}
   (pj/pose :X)
@@ -597,7 +597,7 @@
 
 
 (deftest
- t99_l578
+ t99_l584
  (is
   ((fn
     [pose]
@@ -605,11 +605,11 @@
      (= 2 (count (:poses pose)))
      (= {:x :X, :y :Y} (:mapping (first (:poses pose))))
      (= {:x :X, :y :Z} (:mapping (second (:poses pose))))))
-   v98_l573)))
+   v98_l579)))
 
 
 (def
- v101_l590
+ v101_l596
  (->
   {:X [1 2 3 4 5], :Y [1 2 3 4 5], :Z [1 4 9 16 25]}
   (pj/lay-point :X :Y)
@@ -617,14 +617,14 @@
 
 
 (deftest
- t102_l594
+ t102_l600
  (is
   ((fn [pose] (and (nil? (:poses pose)) (= 2 (count (:layers pose)))))
-   v101_l590)))
+   v101_l596)))
 
 
 (def
- v104_l611
+ v104_l617
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -633,7 +633,7 @@
 
 
 (deftest
- t105_l616
+ t105_l622
  (is
   ((fn
     [pose]
@@ -643,11 +643,11 @@
       {:x :sepal-length, :y :petal-length}
       (:mapping (nth (:poses pose) 2)))
      (= 1 (count (:layers (nth (:poses pose) 2))))))
-   v104_l611)))
+   v104_l617)))
 
 
 (def
- v107_l631
+ v107_l637
  (let
   [via-lay
    (->
@@ -666,14 +666,14 @@
 
 
 (deftest
- t108_l642
+ t108_l648
  (is
   ((fn [m] (= (:via-lay-mappings m) (:via-pose-mappings m)))
-   v107_l631)))
+   v107_l637)))
 
 
 (def
- v110_l654
+ v110_l660
  (try
   (->
    iris
@@ -683,7 +683,7 @@
 
 
 (deftest
- t111_l661
+ t111_l667
  (is
   ((fn
     [msg]
@@ -691,11 +691,11 @@
      (string? msg)
      (re-find #"doesn't exist in the data" msg)
      (re-find #"new sub-pose" msg)))
-   v110_l654)))
+   v110_l660)))
 
 
 (def
- v113_l670
+ v113_l676
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -706,32 +706,32 @@
 
 
 (deftest
- t114_l675
+ t114_l681
  (is
   ((fn
     [fr]
     (and
      (= 2 (count (:poses fr)))
      (= {:x :foo, :y :bar} (:mapping (second (:poses fr))))))
-   v113_l670)))
+   v113_l676)))
 
 
-(def v116_l690 (def tiny {:a [1 2 3 4 5], :b [2 4 3 5 4]}))
+(def v116_l696 (def tiny {:a [1 2 3 4 5], :b [2 4 3 5 4]}))
 
 
-(def v117_l694 (-> tiny (pj/lay-point :a :b)))
-
-
-(deftest
- t118_l697
- (is ((fn [v] (= 5 (:points (pj/svg-summary v)))) v117_l694)))
-
-
-(def v120_l701 (-> tiny (pj/pose :a :b) pj/lay-point pose-summary))
+(def v117_l700 (-> tiny (pj/lay-point :a :b)))
 
 
 (deftest
- t121_l706
+ t118_l703
+ (is ((fn [v] (= 5 (:points (pj/svg-summary v)))) v117_l700)))
+
+
+(def v120_l707 (-> tiny (pj/pose :a :b) pj/lay-point pose-summary))
+
+
+(deftest
+ t121_l712
  (is
   ((fn
     [pose]
@@ -739,11 +739,11 @@
      (= {:x :a, :y :b} (:mapping pose))
      (= 1 (count (:layers pose)))
      (not (contains? pose :poses))))
-   v120_l701)))
+   v120_l707)))
 
 
 (def
- v123_l723
+ v123_l729
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -753,7 +753,7 @@
 
 
 (deftest
- t124_l729
+ t124_l735
  (is
   ((fn
     [v]
@@ -770,11 +770,11 @@
       (contains? (:colors s) "rgb(55,126,184)")
       (contains? (:colors s) "rgb(230,85,13)")
       (= #{"sepal length" "sepal width"} (set axis-titles)))))
-   v123_l723)))
+   v123_l729)))
 
 
 (def
- v126_l749
+ v126_l755
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -785,7 +785,7 @@
 
 
 (deftest
- t127_l756
+ t127_l762
  (is
   ((fn
     [pose]
@@ -796,11 +796,11 @@
       [{:color "#377eb8", :x :sepal-length, :y :sepal-width}
        {:color "#e6550d", :x :petal-length, :y :petal-width}]
       (mapv :mapping (:layers pose)))))
-   v126_l749)))
+   v126_l755)))
 
 
 (def
- v129_l769
+ v129_l775
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -811,7 +811,7 @@
 
 
 (deftest
- t130_l776
+ t130_l782
  (is
   ((fn
     [v]
@@ -823,11 +823,11 @@
       (=
        #{"rgb(55,126,184)" "rgb(230,85,13)" "rgb(77,175,74)"}
        (disj (:colors s) "none")))))
-   v129_l769)))
+   v129_l775)))
 
 
 (def
- v132_l789
+ v132_l795
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -839,7 +839,7 @@
 
 
 (deftest
- t133_l797
+ t133_l803
  (is
   ((fn
     [pose]
@@ -852,11 +852,11 @@
       [{:color "#e6550d"}
        {:color "#4daf4a", :x :sepal-width, :y :petal-width}]
       (mapv :mapping (:layers (second (:poses pose)))))))
-   v132_l789)))
+   v132_l795)))
 
 
 (def
- v135_l810
+ v135_l816
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -869,18 +869,18 @@
 
 
 (deftest
- t136_l816
+ t136_l822
  (is
   ((fn
     [v]
     (let
      [s (pj/svg-summary v)]
      (and (= 2 (:panels s)) (= 300 (:points s)))))
-   v135_l810)))
+   v135_l816)))
 
 
 (def
- v138_l823
+ v138_l829
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -891,18 +891,18 @@
 
 
 (deftest
- t139_l830
+ t139_l836
  (is
   ((fn
     [v]
     (let
      [s (pj/svg-summary v)]
      (and (= 2 (:panels s)) (= 300 (:points s)))))
-   v138_l823)))
+   v138_l829)))
 
 
 (def
- v141_l841
+ v141_l847
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -914,7 +914,7 @@
 
 
 (deftest
- t142_l849
+ t142_l855
  (is
   ((fn
     [pose]
@@ -924,35 +924,38 @@
      (=
       {:x :petal-length, :y :petal-width}
       (:mapping (second (:poses pose))))))
-   v141_l841)))
+   v141_l847)))
 
 
 (def
- v144_l878
+ v144_l888
  (->
-  {:height [1 2 3], :weight [4 5 6], :species ["a" "b" "a"]}
+  (tc/dataset
+   {:height [1 2 3], :weight [4 5 6], :species ["a" "b" "a"]})
   pj/lay-point))
 
 
 (deftest
- t145_l881
- (is ((fn [v] (= 3 (:points (pj/svg-summary v)))) v144_l878)))
+ t145_l891
+ (is ((fn [v] (= 3 (:points (pj/svg-summary v)))) v144_l888)))
 
 
 (def
- v147_l886
+ v147_l896
  (try
-  (-> {:a [1 2], :b [3 4], :c [5 6], :d [7 8]} pj/lay-point)
+  (->
+   (tc/dataset {:a [1 2], :b [3 4], :c [5 6], :d [7 8]})
+   pj/lay-point)
   (catch Exception e (ex-message e))))
 
 
 (deftest
- t148_l892
- (is ((fn [msg] (re-find #"Cannot auto-infer columns" msg)) v147_l886)))
+ t148_l902
+ (is ((fn [msg] (re-find #"Cannot auto-infer columns" msg)) v147_l896)))
 
 
 (def
- v150_l913
+ v150_l923
  (def
   s1-composite
   (pj/pose
@@ -965,11 +968,11 @@
     :data iris})))
 
 
-(def v151_l922 s1-composite)
+(def v151_l932 s1-composite)
 
 
 (deftest
- t152_l924
+ t152_l934
  (is
   ((fn
     [pose]
@@ -981,11 +984,11 @@
      (every?
       (fn [pp] (= 3 (count (:groups (first (:layers (first pp)))))))
       panels)))
-   v151_l922)))
+   v151_l932)))
 
 
 (def
- v154_l936
+ v154_l946
  (def
   s1-siblings
   (pj/pose
@@ -997,11 +1000,11 @@
     :data iris})))
 
 
-(def v155_l944 s1-siblings)
+(def v155_l954 s1-siblings)
 
 
 (deftest
- t156_l946
+ t156_l956
  (is
   ((fn
     [pose]
@@ -1016,11 +1019,11 @@
          (:groups (first (:layers (first (-> sp :plan :panels)))))))
        sub-plots)]
      (= [1 3] panel-groups)))
-   v155_l944)))
+   v155_l954)))
 
 
 (def
- v158_l964
+ v158_l974
  (def
   s2-tree
   (pj/pose
@@ -1033,11 +1036,11 @@
     :data iris})))
 
 
-(def v159_l973 s2-tree)
+(def v159_l983 s2-tree)
 
 
 (deftest
- t160_l975
+ t160_l985
  (is
   ((fn
     [pose]
@@ -1061,11 +1064,11 @@
          count))
        sub-plots)]
      (= [150 3] counts)))
-   v159_l973)))
+   v159_l983)))
 
 
 (def
- v162_l992
+ v162_l1002
  (->
   iris
   (pj/pose :sepal-length :sepal-width {:color :species})
@@ -1074,18 +1077,18 @@
 
 
 (deftest
- t163_l997
+ t163_l1007
  (is
   ((fn
     [v]
     (let
      [s (pj/svg-summary v)]
      (and (= 150 (:points s)) (= 1 (:lines s)))))
-   v162_l992)))
+   v162_l1002)))
 
 
 (def
- v165_l1011
+ v165_l1021
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1094,18 +1097,18 @@
 
 
 (deftest
- t166_l1016
+ t166_l1026
  (is
   ((fn
     [v]
     (let
      [s (pj/svg-summary v)]
      (and (= 150 (:points s)) (= 1 (:lines s)))))
-   v165_l1011)))
+   v165_l1021)))
 
 
 (def
- v168_l1037
+ v168_l1047
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1114,12 +1117,12 @@
 
 
 (deftest
- t169_l1042
- (is ((fn [pose] (= "Iris" (get-in pose [:opts :title]))) v168_l1037)))
+ t169_l1052
+ (is ((fn [pose] (= "Iris" (get-in pose [:opts :title]))) v168_l1047)))
 
 
 (def
- v171_l1047
+ v171_l1057
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1129,18 +1132,18 @@
 
 
 (deftest
- t172_l1053
+ t172_l1063
  (is
   ((fn
     [pose]
     (and
      (= "Two" (get-in pose [:opts :title]))
      (= "Sub" (get-in pose [:opts :subtitle]))))
-   v171_l1047)))
+   v171_l1057)))
 
 
 (def
- v174_l1072
+ v174_l1082
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1150,18 +1153,18 @@
 
 
 (deftest
- t175_l1078
+ t175_l1088
  (is
   ((fn
     [pose]
     (and
      (= {:type :log} (get-in pose [:mapping :x :scale]))
      (= :flip (get-in pose [:opts :coord]))))
-   v174_l1072)))
+   v174_l1082)))
 
 
 (def
- v177_l1088
+ v177_l1098
  (->
   iris
   (pj/pose :sepal-length :sepal-width {:size :petal-length})
@@ -1170,18 +1173,18 @@
 
 
 (deftest
- t178_l1093
+ t178_l1103
  (is
   ((fn
     [pose]
     (=
      {:from :petal-length, :scale {:type :log}}
      (get-in pose [:mapping :size])))
-   v177_l1088)))
+   v177_l1098)))
 
 
 (def
- v180_l1101
+ v180_l1111
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1190,7 +1193,7 @@
 
 
 (deftest
- t181_l1106
+ t181_l1116
  (is
   ((fn
     [pose]
@@ -1199,11 +1202,11 @@
      (=
       {:type :log}
       (-> pose pj/plan :panels first :layers first :size-scale))))
-   v180_l1101)))
+   v180_l1111)))
 
 
 (def
- v183_l1119
+ v183_l1129
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1212,14 +1215,14 @@
 
 
 (deftest
- t184_l1124
+ t184_l1134
  (is
   ((fn [pose] (= :species (get-in pose [:opts :facet-col])))
-   v183_l1119)))
+   v183_l1129)))
 
 
 (def
- v186_l1129
+ v186_l1139
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1228,18 +1231,18 @@
 
 
 (deftest
- t187_l1134
+ t187_l1144
  (is
   ((fn
     [pose]
     (and
      (= :species (get-in pose [:opts :facet-col]))
      (= :species (get-in pose [:opts :facet-row]))))
-   v186_l1129)))
+   v186_l1139)))
 
 
 (def
- v189_l1148
+ v189_l1158
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1248,7 +1251,7 @@
 
 
 (deftest
- t190_l1153
+ t190_l1163
  (is
   ((fn
     [pose]
@@ -1258,15 +1261,15 @@
       rule
       (some
        (fn*
-        [p1__78074#]
-        (when (= :rule-h (:layer-type p1__78074#)) p1__78074#))
+        [p1__11195#]
+        (when (= :rule-h (:layer-type p1__11195#)) p1__11195#))
        layers)]
      (and (some? rule) (= 3.0 (get-in rule [:mapping :y-intercept])))))
-   v189_l1148)))
+   v189_l1158)))
 
 
 (def
- v192_l1163
+ v192_l1173
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1275,7 +1278,7 @@
 
 
 (deftest
- t193_l1168
+ t193_l1178
  (is
   ((fn
     [pose]
@@ -1286,11 +1289,11 @@
      (=
       :rule-h
       (:layer-type (first (:layers (first (:poses pose))))))))
-   v192_l1163)))
+   v192_l1173)))
 
 
 (def
- v195_l1189
+ v195_l1199
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1300,7 +1303,7 @@
 
 
 (deftest
- t196_l1196
+ t196_l1206
  (is
   ((fn
     [pose]
@@ -1312,11 +1315,11 @@
        (fn [sp] (count (:layers (first (-> sp :plan :panels)))))
        (:sub-plots plan))]
      (= [2 1] panel-layer-counts)))
-   v195_l1189)))
+   v195_l1199)))
 
 
 (def
- v198_l1213
+ v198_l1223
  (->
   iris
   (pj/pose :sepal-length :sepal-width {:color :species})
@@ -1324,7 +1327,7 @@
 
 
 (deftest
- t199_l1217
+ t199_l1227
  (is
   ((fn
     [_]
@@ -1347,11 +1350,11 @@
         (= :species (:color d))
         (= :point (:mark d))
         (= 150 (tc/row-count (:data d))))))))
-   v198_l1213)))
+   v198_l1223)))
 
 
 (def
- v201_l1246
+ v201_l1256
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1360,18 +1363,18 @@
 
 
 (deftest
- t202_l1251
+ t202_l1261
  (is
   ((fn
     [pose]
     (let
      [plan (pj/plan pose)]
      (and (:composite? plan) (= 2 (count (:sub-plots plan))))))
-   v201_l1246)))
+   v201_l1256)))
 
 
 (def
- v204_l1264
+ v204_l1274
  (->
   iris
   (pj/pose :sepal-length :sepal-width {:color :species})
@@ -1380,18 +1383,18 @@
 
 
 (deftest
- t205_l1269
+ t205_l1279
  (is
   ((fn
     [pose]
     (let
      [plan (pj/plan pose) panel (first (:panels plan))]
      (and (= 1 (count (:panels plan))) (= 2 (count (:layers panel))))))
-   v204_l1264)))
+   v204_l1274)))
 
 
 (def
- v207_l1282
+ v207_l1292
  (->
   iris
   (pj/pose :sepal-length :sepal-width)
@@ -1400,12 +1403,12 @@
 
 
 (deftest
- t208_l1287
- (is ((fn [pose] (= 3 (count (:panels (pj/plan pose))))) v207_l1282)))
+ t208_l1297
+ (is ((fn [pose] (= 3 (count (:panels (pj/plan pose))))) v207_l1292)))
 
 
 (def
- v210_l1307
+ v210_l1317
  (def
   l4-shared
   (pj/arrange
@@ -1414,11 +1417,11 @@
    {:share-scales #{:x}})))
 
 
-(def v211_l1313 l4-shared)
+(def v211_l1323 l4-shared)
 
 
 (deftest
- t212_l1315
+ t212_l1325
  (is
   ((fn
     [pose]
@@ -1428,8 +1431,8 @@
        [p]
        (mapv
         (fn*
-         [p1__78075#]
-         (get-in p1__78075# [:plan :panels 0 :x-domain]))
+         [p1__11196#]
+         (get-in p1__11196# [:plan :panels 0 :x-domain]))
         (:sub-plots (pj/plan p))))
       domains
       (x-domains pose)
@@ -1440,13 +1443,13 @@
         [(->
           iris
           (tc/select-rows
-           (fn* [p1__78076#] (= "setosa" (:species p1__78076#))))
+           (fn* [p1__11197#] (= "setosa" (:species p1__11197#))))
           (pj/pose :sepal-length :sepal-width)
           pj/lay-point)
          (->
           iris
           (tc/select-rows
-           (fn* [p1__78077#] (= "virginica" (:species p1__78077#))))
+           (fn* [p1__11198#] (= "virginica" (:species p1__11198#))))
           (pj/pose :sepal-length :petal-width)
           pj/lay-point)]
         (if share {:share-scales #{:x}} {})))]
@@ -1456,11 +1459,11 @@
       (apply = domains)
       (apply = (x-domains (cells true)))
       (apply not= (x-domains (cells false))))))
-   v211_l1313)))
+   v211_l1323)))
 
 
 (def
- v214_l1375
+ v214_l1385
  (->
   iris
   (pj/pose
@@ -1469,7 +1472,7 @@
 
 
 (deftest
- t215_l1380
+ t215_l1390
  (is
   ((fn
     [pose]
@@ -1478,7 +1481,7 @@
      (= #{:y :x} (get-in pose [:opts :share-scales]))
      (= 2 (count (:poses pose)))
      (every?
-      (fn* [p1__78078#] (= 2 (count (:poses p1__78078#))))
+      (fn* [p1__11199#] (= 2 (count (:poses p1__11199#))))
       (:poses pose))
      (= {:color :species} (:mapping pose))
      (=
@@ -1487,14 +1490,14 @@
        [{:x :sepal-length, :y :petal-width}
         {:x :sepal-width, :y :petal-width}]]
       (mapv (fn [row] (mapv :mapping (:poses row))) (:poses pose)))))
-   v214_l1375)))
+   v214_l1385)))
 
 
-(def v217_l1409 (pj/cross [:a :b] [:c :d]))
+(def v217_l1419 (pj/cross [:a :b] [:c :d]))
 
 
 (deftest
- t218_l1411
+ t218_l1421
  (is
   ((fn [pairs] (= [[:a :c] [:a :d] [:b :c] [:b :d]] pairs))
-   v217_l1409)))
+   v217_l1419)))